fix(migration): cast step_rate to numeric so ROUND(numeric,int) resolves correctly
PostgreSQL has no ROUND(double precision, integer) overload; ::float caused
SQLSTATE 42883. Changed step_rate CTE column to ::numeric.

+++ supabase/migrations/012_event_spine.sql
@@ -176,7 +176,7 @@ LANGUAGE sql STABLE SECURITY DEFINER AS $$
       COUNT(DISTINCT CASE WHEN event_type = 'mission_resumed'   THEN user_id END) AS returners,
       COUNT(DISTINCT CASE WHEN event_type = 'reward_claimed'    THEN user_id END) AS claimers,
       -- Step completion rate: completed steps / started steps
-      COUNT(*) FILTER (WHERE event_type = 'step_completed')::float /
+      COUNT(*) FILTER (WHERE event_type = 'step_completed')::numeric /
         NULLIF(COUNT(*) FILTER (WHERE event_type = 'step_started'), 0) AS step_rate
     FROM public.mission_events
     WHERE mission_id = p_mission_id
